fix: Refactor utilites to test_utils.go, remove redundant filepath.Clean
