+++ src/internal/handle_file_operation_test.go
@@ -3,7 +3,6 @@ package internal
 import (
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 	"time"
 
@@ -165,149 +164,6 @@ func TestCompressSelectedFiles(t *testing.T) {
 		require.NoError(t, err)
 		assert.Empty(t, entries)
 	})
-}
-
-// Helper function to find item index in panel by name
-func findItemIndexInPanel(panel *filePanel, itemName string) int {
-	for i, elem := range panel.element {
-		if elem.name == itemName {
-			return i
-		}
-	}
-	return -1
-}
-
-// Helper function to setup panel mode and selection
-func setupPanelModeAndSelection(t *testing.T, m *model, useSelectMode bool, itemName string, selectedItems []string) {
-	t.Helper()
-	panel := m.getFocusedFilePanel()
-
-	if useSelectMode {
-		// Switch to select mode and set selected items
-		m.changeFilePanelMode()
-		require.Equal(t, selectMode, panel.panelMode)
-		panel.selected = selectedItems
-	} else {
-		// Find the item in browser mode
-		itemIndex := findItemIndexInPanel(panel, itemName)
-		require.NotEqual(t, -1, itemIndex, "%s should be found in panel", itemName)
-		panel.cursor = itemIndex
-	}
-}
-
-// Helper function to perform copy or cut operation
-func performCopyOrCutOperation(t *testing.T, m *model, isCut bool) {
-	t.Helper()
-	if isCut {
-		TeaUpdateWithErrCheck(t, m, utils.TeaRuneKeyMsg(common.Hotkeys.CutItems[0]))
-	} else {
-		TeaUpdateWithErrCheck(t, m, utils.TeaRuneKeyMsg(common.Hotkeys.CopyItems[0]))
-	}
-}
-
-// Helper function to verify clipboard state after copy/cut
-func verifyClipboardState(t *testing.T, m *model, isCut bool, useSelectMode bool, selectedItemsCount int) {
-	t.Helper()
-	assert.Equal(t, isCut, m.copyItems.cut, "Clipboard cut state should match operation")
-	if useSelectMode {
-		assert.Len(t, m.copyItems.items, selectedItemsCount, "Clipboard should contain all selected items")
-	} else {
-		assert.Len(t, m.copyItems.items, 1, "Clipboard should contain one item")
-	}
-}
-
-// Helper function to navigate to target directory if different from start
-func navigateToTargetDir(t *testing.T, m *model, startDir, targetDir string) {
-	t.Helper()
-	if targetDir != startDir {
-		m.updateCurrentFilePanelDir(targetDir)
-		TeaUpdateWithErrCheck(t, m, nil)
-	}
-}
-
-// Helper function to get original path for existence check
-func getOriginalPath(useSelectMode bool, itemName, startDir string) string {
-	if !useSelectMode && itemName != "" {
-		return filepath.Join(startDir, itemName)
-	}
-	return ""
-}
-
-// Helper function to verify file or directory exists
-func verifyPathExists(t *testing.T, path, message string) {
-	t.Helper()
-	if strings.Contains(path, ".txt") {
-		assert.FileExists(t, path, message)
-	} else {
-		assert.DirExists(t, path, message)
-	}
-}
-
-// Helper function to verify file or directory doesn't exist after cut
-func verifyPathNotExistsEventually(t *testing.T, path, message string) {
-	t.Helper()
-	assert.Eventually(t, func() bool {
-		_, err := os.Stat(path)
-		return os.IsNotExist(err)
-	}, time.Second, 10*time.Millisecond, message)
-}
-
-// Helper function to verify expected destination files exist
-func verifyDestinationFiles(t *testing.T, targetDir string, expectedDestFiles []string) {
-	t.Helper()
-	for _, expectedFile := range expectedDestFiles {
-		destPath := filepath.Join(targetDir, expectedFile)
-		assert.Eventually(t, func() bool {
-			_, err := os.Stat(destPath)
-			return err == nil
-		}, time.Second, 10*time.Millisecond, "%s should exist in destination", expectedFile)
-	}
-}
-
-// Helper function to verify prevented paste results
-func verifyPreventedPasteResults(t *testing.T, m *model, originalPath string) {
-	t.Helper()
-	if originalPath != "" {
-		verifyPathExists(t, originalPath, "Original file should still exist when paste is prevented")
-	}
-	// Clipboard should not be cleared when paste is prevented
-	assert.NotEmpty(t, m.copyItems.items, "Clipboard should not be cleared when paste is prevented")
-}
-
-// Helper function to verify successful paste results
-func verifySuccessfulPasteResults(t *testing.T, targetDir string, expectedDestFiles []string, originalPath string, shouldOriginalExist bool) {
-	t.Helper()
-	// Verify expected files were created in destination
-	verifyDestinationFiles(t, targetDir, expectedDestFiles)
-
-	// Verify original file existence based on operation type
-	if originalPath != "" {
-		if shouldOriginalExist {
-			verifyPathExists(t, originalPath, "Original file should exist after copy operation")
-		} else {
-			verifyPathNotExistsEventually(t, originalPath, "Original file should not exist after cut operation")
-		}
-	}
-
-	// TODO: Need to add a test to verify clipboard state.
-}
-
-// Helper function to setup model and perform copy/cut operation
-func setupModelAndPerformOperation(t *testing.T, startDir string, useSelectMode bool, itemName string, selectedItems []string, isCut bool) *model {
-	t.Helper()
-	m := defaultTestModel(startDir)
-	TeaUpdateWithErrCheck(t, &m, nil)
-
-	setupPanelModeAndSelection(t, &m, useSelectMode, itemName, selectedItems)
-	performCopyOrCutOperation(t, &m, isCut)
-
-	selectedItemsCount := len(selectedItems)
-	if !useSelectMode {
-		selectedItemsCount = 1
-	}
-	verifyClipboardState(t, &m, isCut, useSelectMode, selectedItemsCount)
-
-	return &m
 }
 
 func TestPasteItem(t *testing.T) {
@@ -485,4 +341,24 @@ func TestPasteItem(t *testing.T) {
 		// Verify duplicate file with different name
 		verifyDestinationFiles(t, destDir, []string{"duplicate(1).txt"})
 	})
+}
+
+// ------  Very specific utilities that are required for this test case file only
+
+// Helper function to setup model and perform copy/cut operation
+func setupModelAndPerformOperation(t *testing.T, startDir string, useSelectMode bool, itemName string, selectedItems []string, isCut bool) *model {
+	t.Helper()
+	m := defaultTestModel(startDir)
+	TeaUpdateWithErrCheck(t, &m, nil)
+
+	setupPanelModeAndSelection(t, &m, useSelectMode, itemName, selectedItems)
+	performCopyOrCutOperation(t, &m, isCut)
+
+	selectedItemsCount := len(selectedItems)
+	if !useSelectMode {
+		selectedItemsCount = 1
+	}
+	verifyClipboardState(t, &m, isCut, useSelectMode, selectedItemsCount)
+
+	return &m
 }
